requirements.txt
Flask==1.1.1
h5py==2.10.0
Keras==2.3.1
matplotlib==3.1.2
numpy==1.19.2
pandas==0.25.3
plotly==4.10.0
scikit-learn==0.23.2
streamlit==0.66.0
tensorflow-cpu==2.2.0
scipy==1.4.1